Move Windows implementation to a new package
