+++ window-styler/src/jvmMain/kotlin/com/mayakapps/compose/windowstyler/WindowStyle.kt
@@ -4,13 +4,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.window.WindowScope
+import com.mayakapps.compose.windowstyler.windows.WindowsWindowManager
 
 @Composable
 fun WindowScope.WindowStyle(
     isDarkTheme: Boolean = false,
     backdropType: WindowBackdrop = WindowBackdrop.Default,
 ) {
-    val manager = remember { WindowManager(window, isDarkTheme, backdropType) }
+    val manager = remember { WindowsWindowManager(window, isDarkTheme, backdropType) }
 
     LaunchedEffect(isDarkTheme) {
         manager.isDarkTheme = isDarkTheme
