+++ window-styler/src/jvmMain/kotlin/com/mayakapps/compose/windowstyler/windows/jna/User32.kt
@@ -1,6 +1,6 @@
-package com.mayakapps.compose.windowstyler.jna
+package com.mayakapps.compose.windowstyler.windows.jna
 
-import com.mayakapps.compose.windowstyler.jna.structs.WindowCompositionAttributeData
+import com.mayakapps.compose.windowstyler.windows.jna.structs.WindowCompositionAttributeData
 import com.sun.jna.Native
 import com.sun.jna.platform.win32.WinDef
 import com.sun.jna.win32.StdCallLibrary
