+++ window-styler/src/jvmMain/kotlin/com/mayakapps/compose/windowstyler/windows/jna/enums/AccentState.kt
@@ -1,4 +1,4 @@
-package com.mayakapps.compose.windowstyler.jna.enums
+package com.mayakapps.compose.windowstyler.windows.jna.enums
 
 @Suppress("SpellCheckingInspection", "unused")
 internal enum class AccentState(val value: Int) {
